refactor(#188): move managers to a new files
