+++ src/modules/ContentScript/PageTranslationContext.ts
@@ -9,27 +9,14 @@ import { getSitePreferences } from '../../requests/backend/autoTranslation/siteP
 import { getLanguagePreferences } from '../../requests/backend/autoTranslation/languagePreferences/getLanguagePreferences';
 import { isRequireTranslateBySitePreferences } from '../../layouts/PageTranslator/PageTranslator.utils/utils';
 
-import { PageTranslator } from '../PageTranslator/PageTranslator';
-import {
-	Options as SelectTranslatorOptions,
-	SelectTranslator,
-} from '../SelectTranslator';
+import { SelectTranslatorManager } from '../SelectTranslator/SelectTranslatorManager';
+import { PageTranslatorManager } from '../PageTranslator/PageTranslatorManager';
 
-const buildSelectTranslatorOptions = (
-	{ mode, ...options }: AppConfigType['selectTranslator'],
-	{ pageLanguage }: { pageLanguage?: string },
-): SelectTranslatorOptions => ({
-	...options,
-	pageLanguage,
-	quickTranslate: mode === 'quickTranslate',
-	enableTranslateFromContextMenu: mode === 'contextMenu',
-});
-
-type PageData = {
+export type PageData = {
 	language: string | null;
 };
 
-type PageTranslationOptions = {
+export type PageTranslationOptions = {
 	from: string;
 	to: string;
 };
@@ -38,132 +25,6 @@ type TranslatorsState = {
 	pageTranslation: PageTranslationOptions | null;
 	textTranslation: boolean;
 };
-
-// TODO: move to a file
-class SelectTranslatorManager {
-	private $state;
-	constructor(
-		$state: Store<{
-			enabled: boolean;
-			config: AppConfigType['selectTranslator'];
-			pageData: PageData;
-		}>,
-	) {
-		this.$state = $state;
-	}
-
-	private selectTranslator: SelectTranslator | null = null;
-
-	public getSelectTranslator() {
-		return this.selectTranslator;
-	}
-
-	public start() {
-		// Manage text translation instance
-		this.$state.watch(({ config: preferences, pageData }) => {
-			console.warn('TT prefs', preferences);
-
-			if (preferences.enabled) {
-				const pageLanguage = pageData.language || undefined;
-				const config = buildSelectTranslatorOptions(preferences, {
-					pageLanguage,
-				});
-
-				if (this.selectTranslator === null) {
-					this.selectTranslator = new SelectTranslator(config);
-				} else {
-					const isRun = this.selectTranslator.isRun();
-					if (isRun) {
-						this.selectTranslator.stop();
-					}
-
-					this.selectTranslator = new SelectTranslator(config);
-
-					if (isRun) {
-						this.selectTranslator.start();
-					}
-				}
-			} else {
-				if (this.selectTranslator === null) return;
-
-				if (this.selectTranslator.isRun()) {
-					this.selectTranslator.stop();
-				}
-
-				this.selectTranslator = null;
-			}
-		});
-
-		// Manage text translation state
-		const $isTextTranslationStarted = this.$state.map(({ enabled }) => enabled);
-		$isTextTranslationStarted.watch((isTranslating) => {
-			console.warn('TT state', isTranslating);
-
-			if (this.selectTranslator === null) return;
-			if (isTranslating === this.selectTranslator.isRun()) return;
-
-			if (isTranslating) {
-				this.selectTranslator.start();
-			} else {
-				this.selectTranslator.stop();
-			}
-		});
-	}
-}
-
-// TODO: move to a file
-class PageTranslatorManager {
-	private $state;
-	private pageTranslator: PageTranslator;
-
-	constructor(
-		$state: Store<{
-			state: PageTranslationOptions | null;
-			config: AppConfigType['pageTranslator'];
-		}>,
-	) {
-		this.$state = $state;
-
-		// Create instances for translation
-		const currentState = $state.getState();
-		this.pageTranslator = new PageTranslator(currentState.config);
-	}
-
-	public getDomTranslator() {
-		return this.pageTranslator;
-	}
-
-	public start() {
-		// Manage page translation instance
-		this.$state.watch(({ config }) => {
-			if (!this.pageTranslator.isRun()) {
-				this.pageTranslator.updateConfig(config);
-				return;
-			}
-
-			const direction = this.pageTranslator.getTranslateDirection();
-			if (direction === null) {
-				throw new TypeError('Invalid response from getTranslateDirection method');
-			}
-
-			this.pageTranslator.stop();
-			this.pageTranslator.updateConfig(config);
-			this.pageTranslator.run(direction.from, direction.to);
-		});
-
-		// Manage page translation state
-		this.$state.watch(({ state: pageTranslation }) => {
-			const shouldTranslate = pageTranslation !== null;
-			if (shouldTranslate === this.pageTranslator.isRun()) return;
-
-			if (pageTranslation !== null) {
-				this.pageTranslator.run(pageTranslation.from, pageTranslation.to);
-			} else {
-				this.pageTranslator.stop();
-			}
-		});
-	}
-}
 
 export class PageTranslationContext {
 	private $config: Store<AppConfigType>;
